+++ chapter_05/exercise_5_5/strncat.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <string.h>
 
-void strcat_ptr(char *s, char *t, unsigned int n);
+void strcat_ptr(char *s, char *t, size_t n);
 
 int main(void)
 {
   char s[100] = "This is the first string";
   char *t = ", this second string!";
-  unsigned int nr_chars = 5;
+  size_t nr_chars = 5;
 
   strcat_ptr(s, t, nr_chars);
 
@@ -17,10 +17,10 @@ int main(void)
 }
 
 // Concatenate t to end of s; s must be big enough.
-void strcat_ptr(char *s, char *t, unsigned int n)
+void strcat_ptr(char *s, char *t, size_t n)
 {
   // Find the end of s
-  unsigned int s_length = strlen(s);
+  size_t s_length = strlen(s);
 
   // Move the s pointer to the end of the s string.
   s += s_length;
